Move graphql things out from `components` directory
