+++ frontend/src/api/mutations/DeletePoint.ts
